requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via
    #   label-studio-ml
    #   label-studio-tools
async-timeout==4.0.3
    # via redis
attr==0.3.1
    # via label-studio-ml
attrs==23.2.0
    # via label-studio-ml
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   rq
colorama==0.4.6
    # via label-studio-ml
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dill==0.3.8
    # via -r requirements.in
docutils==0.20.1
    # via python-daemon
exceptiongroup==1.2.0
    # via pytest
ffmpeg-python==0.2.0
    # via -r requirements.in
filelock==3.13.1
    # via
    #   torch
    #   triton
flask==1.1.2
    # via label-studio-ml
fonttools==4.49.0
    # via matplotlib
fsspec==2024.2.0
    # via torch
future==1.0.0
    # via ffmpeg-python
idna==3.6
    # via requests
imageio==2.34.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
itsdangerous==2.0.1
    # via
    #   flask
    #   label-studio-ml
jinja2==3.0.3
    # via
    #   flask
    #   label-studio-ml
    #   torch
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
label-studio-ml==1.0.9
    # via -r requirements.in
label-studio-tools==0.0.3
    # via label-studio-ml
lockfile==0.12.2
    # via python-daemon
luigi==3.5.0
    # via -r requirements.in
lxml==5.1.0
    # via
    #   label-studio-ml
    #   label-studio-tools
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   contourpy
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.18.1
    # via torch
nvidia-nvjitlink-cu12==12.3.101
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
opencv-python==4.9.0.80
    # via
    #   -r requirements.in
    #   ultralytics
packaging==23.2
    # via
    #   matplotlib
    #   pytest
pandas==2.2.1
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
pillow==10.2.0
    # via
    #   imageio
    #   label-studio-ml
    #   matplotlib
    #   torchvision
    #   ultralytics
pluggy==1.4.0
    # via pytest
psutil==5.9.8
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
py4j==0.10.9.7
    # via pyspark
pyarrow==15.0.2
    # via -r requirements.in
pyparsing==3.1.1
    # via matplotlib
pyspark==3.5.1
    # via -r requirements.in
pytest==8.0.2
    # via -r requirements.in
python-daemon==3.0.1
    # via luigi
python-dateutil==2.8.2
    # via
    #   luigi
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via ultralytics
redis==5.0.3
    # via rq
requests==2.31.0
    # via
    #   -r requirements.in
    #   label-studio-ml
    #   torchvision
    #   ultralytics
rq==1.16.1
    # via -r requirements.in
scikit-learn==1.4.1.post1
    # via -r requirements.in
scipy==1.12.0
    # via
    #   scikit-learn
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tenacity==8.2.3
    # via luigi
thop==0.1.1.post2209072238
    # via ultralytics
threadpoolctl==3.4.0
    # via scikit-learn
tomli==2.0.1
    # via pytest
torch==2.1.2
    # via
    #   thop
    #   torchvision
    #   ultralytics
torchvision==0.16.2
    # via
    #   -r requirements.in
    #   ultralytics
tornado==6.4
    # via luigi
tqdm==4.66.2
    # via
    #   -r requirements.in
    #   ultralytics
triton==2.1.0
    # via torch
typing-extensions==4.10.0
    # via torch
tzdata==2024.1
    # via pandas
ultralytics==8.1.19
    # via -r requirements.in
urllib3==2.2.1
    # via requests
werkzeug==2.0.2
    # via
    #   flask
    #   label-studio-ml

# The following packages are considered to be unsafe in a requirements file:
# setuptools




luigi>=3.3.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.2.0
ultralytics>=8.0.0
